+++ NEWS.md
@@ -10,6 +10,7 @@ None
 4. Fix #68 - Add `hai_hyperbolic_vec()` function.
 5. Fix #70 - Add `hai_hyperbolic_augment()` function.
 6. Fix #75 - Add `hai_fourier_vec()`,`hai_fourier_augment()`,`step_hai_fourier()` functions.
+7. Fix #89 - Add `hai_fourier_discrete_vec()`, `hai_fourier_discrete_augment()`, `step_hai_fourier_discrete()`
 
 ## Minor Fixes and Improvments
 1. Fix #81 - Add process to register s3 methods to work with `tune` objects.